% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_objects.R
\name{buildForest}
\alias{buildForest}
\title{Building model objects functions}
\usage{
buildForest(
  treeData = NULL,
  shrubData = NULL,
  seedData = NULL,
  miscData = NULL
)
}
\arguments{
\item{treeData}{tree data dataframe}

\item{shrubData}{shrub data dataframe}

\item{seedData}{seed bank data dataframe}

\item{miscData}{miscelaneous data dataframe}
}
\description{
This functions create the objects needed for the model runs
}
\concept{ModelObjects}
