% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_data_path_functions.R
\name{data_path}
\alias{data_path}
\alias{set_data_path}
\alias{get_data_path}
\title{Site data path}
\usage{
set_data_path(path)

get_data_path()
}
\arguments{
\item{path}{}
}
\description{
Functions to get/set path to site data
}
