% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{run_reports}
\alias{run_reports}
\title{Main function to run reports}
\usage{
run_reports(type = "evaluation", sites, model, confs, wd)
}
\arguments{
\item{type}{which kind of report must be generated, "evaluation"}

\item{sites}{A character vector with the sites codes}

\item{model}{A string with the model to be run ("spwb" or "growth")}

\item{confs}{A character vector with names of configuration (control) settings}

\item{wd}{Complete path to the validation directory}
}
\description{
Runs reports for a set of sites under a set of configurations
}
